divide and conquer
