+++ app/models/carto/helpers/varnish_cache_handler.rb
@@ -1,4 +1,4 @@
-module VarnishCacheHandler
+module Carto::VarnishCacheHandler
   def invalidate_varnish_cache(options = {})
     options[:regex] ||= '.*'
     CartoDB::Varnish.new.purge("#{database_name}#{options[:regex]}")
